+++ lending/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.2.0] lending
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/lending/v/0.2.0 - lending
+
+## 2023-10-30 01:53:28
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] lending
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/lending/v/0.3.0 - lending